+++ persistence/actor_shared_sql.go
@@ -4,8 +4,7 @@ import (
 	"context"
 	"encoding/hex"
 	"fmt"
-
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/pokt-network/pocket/persistence/types"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/modules"