+++ shared/modules/persistence_module.go
@@ -70,7 +70,7 @@ type PersistenceWriteContext interface {
 	AddPoolAmount(name string, amount string) error
 	SubtractPoolAmount(name string, amount string) error
 	SetPoolAmount(name string, amount string) error
-	InsertPool(name string, address []byte, amount string) error // TODO(#149): remove address from pool
+	InsertPool(name string, amount string) error
 
 	// Account Operations
 	AddAccountAmount(address []byte, amount string) error
@@ -113,11 +113,11 @@ type PersistenceWriteContext interface {
 
 	// Param Operations
 	InitGenesisParams(params *genesis.Params) error
-	SetParam(paramName string, value interface{}) error
+	SetParam(paramName string, value any) error
 
 	// Flag Operations
 	InitFlags() error
-	SetFlag(paramName string, value interface{}, enabled bool) error
+	SetFlag(paramName string, value any, enabled bool) error
 }
 
 type PersistenceReadContext interface {
@@ -127,9 +127,8 @@ type PersistenceReadContext interface {
 
 	// CONSOLIDATE: BlockHash / AppHash / StateHash
 	// Block Queries
-	GetLatestBlockHeight() (uint64, error)         // Returns the height of the latest block in the persistence layer
-	GetBlockHash(height int64) (string, error)     // Returns the app hash corresponding to the height provided
-	GetBlocksPerSession(height int64) (int, error) // TECHDEBT(#286): Deprecate this method
+	GetLatestBlockHeight() (uint64, error)     // Returns the height of the latest block in the persistence layer
+	GetBlockHash(height int64) (string, error) // Returns the app hash corresponding to the height provided
 
 	// Pool Queries
 
@@ -161,7 +160,6 @@ type PersistenceReadContext interface {
 	GetServiceNodePauseHeightIfExists(address []byte, height int64) (int64, error)
 	GetServiceNodeOutputAddress(operator []byte, height int64) (output []byte, err error)
 	GetServiceNodeCount(chain string, height int64) (int, error)
-	GetServiceNodesPerSessionAt(height int64) (int, error) // TECHDEBT(#286): Deprecate this method
 
 	// Fisherman Queries
 	GetAllFishermen(height int64) ([]*coreTypes.Actor, error)
@@ -189,6 +187,7 @@ type PersistenceReadContext interface {
 	GetIntParam(paramName string, height int64) (int, error)
 	GetStringParam(paramName string, height int64) (string, error)
 	GetBytesParam(paramName string, height int64) ([]byte, error)
+	GetParameter(paramName string, height int64) (any, error)
 
 	// Flags
 	GetIntFlag(paramName string, height int64) (int, bool, error)